+++ src/extension.ts
@@ -6,6 +6,7 @@ import { StatusBarManager } from './statusBar';
 import { ModelManager } from './modelManager';
 import { ModelManagerPanel } from './modelManagerPanel';
 import { showPipelineSettings } from './pipelineSettingsUI';
+import { showLanguageSelector } from './languageSelector';
 
 let engine: VoxPilotEngine | undefined;
 let statusBar: StatusBarManager;
@@ -46,6 +47,7 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand('voxpilot.inlineVoiceInput', () => engine?.inlineVoiceInput()),
     vscode.commands.registerCommand('voxpilot.transcriptHistory', () => engine?.showTranscriptHistory()),
     vscode.commands.registerCommand('voxpilot.sendToChat', () => engine?.sendLastToChat()),
+    vscode.commands.registerCommand('voxpilot.selectLanguage', () => engine?.selectLanguage()),
     vscode.commands.registerCommand('voxpilot.clearCache', () => clearCache(context)),
     vscode.commands.registerCommand('voxpilot.pipelineSettings', () => engine ? showPipelineSettings(engine.pipeline) : undefined),
     vscode.commands.registerCommand('voxpilot.modelManager.download', (item) => modelPanel.downloadModel(item)),